Make ChorusSystem more UI independent
This was done by splitting out a UI independent ChorusSystemSimple
base class.
