+++ src/SampleApp/Form1.cs
@@ -3,10 +3,9 @@ using System.Diagnostics;
 using System.IO;
 using System.Windows.Forms;
 using Chorus;
-using Chorus.UI.Notes;
+using Chorus.notes;
 using Chorus.UI.Notes.Browser;
 using Chorus.UI.Review;
-using Chorus.Utilities;
 using Chorus.VcsDrivers;
 using Palaso.IO;
 using SampleApp.Properties;
@@ -59,11 +58,9 @@ namespace SampleApp
 			//Also note that this is not the same name as that used for any given network repository credentials;
 			//Rather, it's the name which will show in the history, and besides Notes that this user makes.
 			_chorusSystem = new ChorusSystem(shoppingListDir);
-			_chorusSystem.DisplaySettings = new ChorusNotesDisplaySettings()
-			{
-				WritingSystemForNoteLabel = new TestWritingSystem("Algerian"),
-				WritingSystemForNoteContent = new TestWritingSystem("Bradley Hand ITC")
-			};
+			_chorusSystem.DisplaySettings.WritingSystemForNoteLabel = new TestWritingSystem("Algerian");
+			_chorusSystem.DisplaySettings.WritingSystemForNoteContent =
+				new TestWritingSystem("Bradley Hand ITC");
 
 			_chorusSystem.Init(userName);
